# Vertex and attribute compression