… doc/classes/EditorSceneFormatImporter.xml
… doc/classes/MeshDataTool.xml
… doc/classes/RenderingServer.xml
… doc/classes/ResourceImporterOBJ.xml
… doc/classes/ResourceImporterScene.xml
… drivers/gles3/storage/mesh_storage.cpp
… modules/gltf/gltf_state.h
… modules/mono/glue/GodotSharp/GodotSharp/Compat.cs
… scene/3d/soft_body_3d.cpp
… scene/3d/soft_body_3d.h
… scene/3d/sprite_3d.cpp
… scene/3d/sprite_3d.h
… scene/resources/immediate_mesh.cpp
… scene/resources/importer_mesh.cpp
… scene/resources/importer_mesh.h
… scene/resources/material.h
… scene/resources/mesh.cpp
… scene/resources/mesh.h
… scene/resources/mesh_data_tool.compat.inc
… scene/resources/mesh_data_tool.cpp
… scene/resources/mesh_data_tool.h
… scene/resources/surface_tool.cpp
… scene/resources/surface_tool.h
… servers/rendering/renderer_rd/forward_clustered/render_forward_clustered.cpp
… servers/rendering/renderer_rd/forward_clustered/render_forward_clustered.h
… servers/rendering/renderer_rd/forward_clustered/scene_shader_forward_clustered.h
… servers/rendering/renderer_rd/forward_mobile/render_forward_mobile.cpp
… servers/rendering/renderer_rd/forward_mobile/render_forward_mobile.h
… servers/rendering/renderer_rd/forward_mobile/scene_shader_forward_mobile.cpp
… servers/rendering/renderer_rd/forward_mobile/scene_shader_forward_mobile.h