This scheme is backwards compatible so old meshes will continue to work.

Users can get a performance boost by reimporting assets

The rationale behind this PR can be found here: https://files.godot.foundation/s/259N9YQEPcBZwMx

In short, the main reason behind making this PR is bandwidth reduction (passing less data around the GPU per draw call). Bandwidth reduction is valuable in two cases:
1. If the scene is bandwidth limited, reducing bandwidth can improve performance
2. Reducing bandwidth reduces power consumption on mobile and power-constrained devices. 

From our perspective, reducing bandwidth also has the added benefit that we can now confirm that we are not generally bandwidth bound (i.e. we aren't boundwidth bound in _all_ scenes) which means we can focus our next performance improvements elsewhere (VGPRs / instructions).

### Usage notes

After merging this PR all meshes will automatically start using a slightly different vertex attribute layout:
```
P = Vertex position
N = normal
T = tangent

// Before this PR
PNTPNTPNTPNT

// After this PR
PPPPNTNTNTNT
```

This is the recommended layout for mobile devices as it means that shadow passes and tiling passes only ever have to read vertex positions and it doesn't hurt performance on Desktop. 

The renderer will automatically use this layout even if the mesh doesn't. To accommodate that, I have added a mesh format version number so we can perform automatic upgrades between versions. This avoids the issue we had last summer when integrating octahedral compression. Three things to note about this though:
1. When running scenes with the old mesh format, users will get warnings for each mesh that needs to be upgraded
2. Meshes need to save the new mesh format to stop getting the warning. In some cases this means they will have to reimport the mesh. Upon saving a mesh though, the new format will be used. 
3. Once the new format is used, users can't go back to an earlier version without re-importing their mesh. 

In addition to the vertex layout change, we also introduced optional compression for vertices, normals, tangents, and UVs. The importers will automatically detect if the mesh is a candidate for compression and apply the appropriate flags. For UVs, we can compress them as long as they are all within the 0-1 range. While this is technically a lossy compression, there won't be a visible difference on textures with a size smaller than 65536 pixels, so it is safe to apply). For vertices we do two things:
1. Scale the vertices into the 0-1 range using the surface's AABB (this allows us to store the meshes in RGB16 instead of RGB32
2. Store Normal (vec3) + Tangent (vec4) as Axis (vec3)+ Angle (float). The Axis replaces the normal and tangent fits into the A channel of the vertex. 

I implemented this compression scheme in a way that does not require additional shader variants, so there will be no effect on shader compile time. However, this means that there are more calculations in the shader, and VGPR usage may be affected on some compilers. 

When the compression flags are not used, the vertex format remains the same as it was before this PR. 

### Performance

So far I have tested on my laptop using an integrated GPU and an NVidia 3050m. There is no difference when using the forward+ renderer, a slight improvement when using the compatibility renderer, and about a 10% improvement when using the mobile renderer. This is likely due to the fact that I had to refactor the mobile renderer to pass the necessary data. 

On the TPS demo on integrated graphics, I am seeing an improvement from ~17 FPS to ~20FPS (Forward+). 
Using the [Synty Vikings demo](https://syntystore.com/products/polygon-vikings-pack) When using the mobile renderer, I get an increase from 40 ms per frame to 37 ms per frame. 

Running the TPS demo on mobile (Pixel 7 and Pixel 1) I see no performance difference (likely because both are instruction bound, not bandwidth bound).

~More performance testing is needed before merging. But so far, in all cases the performance is either the exact same or slightly better.~

Todos:
1. ~Add additional optimization by de-interleaving vertex positions (this won't be backwards compatible and so we need to add version identifiers to mesh data~
2. ~Add a setting to force disable optimization on an imported mesh. Generally, the auto-detection works quite well. But it relies on the mesh having both proper normals and tangents. If the mesh has improper tangents, the result is also bad.~ In the end this setting wasn't needed. The issue came from a scene with non-continuous UVs, so the generated tangents were bogus. The solution was to allow GLTFs to respect the "ensure tangents" setting. Previously tangents were always generated if not supplied. 
3. ~more performance testing with heavier scenes on mobile devices (especially scenes with more shadows).~