@clayjohn 2023-09-05
Thanks for the reviews AThousandShips!

Just pushed another version with docs and fixes for CI. This PR should be ready to go now. I don't expect to find any performance regressions in my final testing. 

One catch, I am away on holidays from September 7 to 19. So if we merge this anytime soon I won't be around to jump on bugs. So we may want to merge this the week of the 18th so that I can be available to fix any bugs that arise. While I have done my best to test extensively, it is inevitable that bugs will appear from such a large change to a fundamental system