@akien-mga 2023-09-05
Changing `gltf_document.cpp` caused it to be recompiled, and a new MSVC warning was found (not caused by this PR but could be fixed here too to pass CI):
```
Error: modules\gltf\gltf_document.cpp(2958): error C2220: the following warning is treated as an error
Warning: modules\gltf\gltf_document.cpp(2958): warning C4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?)
```
I think this kind of error can be fixed by changing `1 << l` to `1U << l`.