@dsnopek 2023-09-07
I didn't test and I only looked at the GDExtension compatibility code, but those bits look good to me!