@clayjohn 2023-09-27
> > While this is technically a lossy compression, there won't be a visible difference on textures with a size smaller than 65536 pixels, so it is safe to apply
> 
> you mean x and y size or total pixel count?

x and y size