@raulsntos 2023-09-27
> The change of parameter type to `SurfaceTool.commit()` requires a compat binding for GDExtension, but it seems to generate a conflicting compat method for C#

I don't know what's going on here, the method list contains the method `commit` twice (same hash and everything). Actually, could it be that we're not considering the metadata when calculating the method's hashes?