@clayjohn 2023-09-27
Rebased and addressed Bastiaan's comment. Maybe the rebase will fix the C# issue?