@raulsntos 2023-09-28
From what I understand, the method hash is calculated from the argument type but it's not taking the type metadata into consideration:

https://github.com/godotengine/godot/blob/4c3dc26367518e006f8555c12f5d2df0b8a28192/core/object/method_bind.cpp#L43

So for `int32` and `int64`, the computed hash is the same. Since the registered method and the compatibility method shared the same hash, the C# bindings generator ends up generating the registered method twice.

I don't know how to change the hash computation without breaking compatibility though.