@clayjohn 2023-09-28
CC @dsnopek Do you think there is a way to make the binder take arg types into account without breaking compatibility? Similar to your recent PR fixing the hash for default arguments?