@reduz 2023-09-28
The binder always uses int64 as arg type, so this is a problem on the C# side.