@reduz 2023-09-28
To clarify you should not  have any problem with old code using this, new and old versions of the C# code generated will continue working.