@dsnopek 2023-09-28
> Do you think there is a way to make the binder take arg types into account without breaking compatibility? Similar to your recent PR fixing the hash for default arguments?

Like Juan says, in GDExtension, all integer arguments are encoded as `int64_t` (ie. cast to `int64_t`, transferred across the boundary, then cast back to correct type). So, for the purposes of GDExtension we do want the hashes to be the same. I think the problem here stems from C# using the same data as GDExtension, while having slightly different requirements.

Could the C# code generator maybe have a list of exceptions that prevent the automatic generation, and then some handwritten code could be added for this case?