@raulsntos 2023-09-28
> To clarify you should not have any problem with old code using this, new and old versions of the C# code generated will continue working.

That's not how compatibility works in C#, if the method doesn't exist in the new bindings it will not be found by old code.

> Could the C# code generator maybe have a list of exceptions that prevent the automatic generation, and then some handwritten code could be added for this case?

We could avoid generating compatibility methods that have the same hash as an existing non-compat method. However, I feel like this defeats the purpose of https://github.com/godotengine/godot/pull/80527. The goal with that PR was automating this so contributors didn't have to add compatibility methods to C# manually (see [Compat.cs](https://github.com/godotengine/godot/blob/4c3dc26367518e006f8555c12f5d2df0b8a28192/modules/mono/glue/GodotSharp/GodotSharp/Compat.cs)).

> Like Juan says, in GDExtension, all integer arguments are encoded as `int64_t` (ie. cast to `int64_t`, transferred across the boundary, then cast back to correct type). So, for the purposes of GDExtension we do want the hashes to be the same.

But if an argument has different type metadata, doesn't that make it different enough to guarantee having different hashes? In fact, this PR is providing a compatibility method precisely because it changes the metadata of an argument that used to be `int32_t` and is now `int64_t`. So if the metadata didn't matter, why do we need to add a compatibility method?

> I think the problem here stems from C# using the same data as GDExtension, while having slightly different requirements.

The problem is that I can't retrieve the compatibility method when it has the same hash as a non-compat method, because `ClassDB::get_method_with_compatibility` will always retrieve the non-compat method. I feel like having multiple methods with the same hash registered in ClassDB is not right.

But I understand that changing the way we compute the method hash is not a solution, since it would also change the hash for every other method in `extension_api.json` and therefore break compatibility everywhere. So I don't know how to solve this, other than going back to adding compat methods manually for C#.