@dsnopek 2023-09-28
> But if an argument has different type metadata, doesn't that make it different enough to guarantee having different hashes?

On a purely technical level, for GDExtension, we don't need for these hashes to be different in this case, since the encoding used when calling the function won't change.

That said, we definitely didn't get the hashing algorithm perfectly correct. :-) There are cases where the hash will change (for example, changing from one enum type to another), but technically, we don't need it to be different because the encoding is the same. It may or may not be worth trying to fix those cases eventually.

> So if the metadata didn't matter, why do we need to add a compatibility method?

I don't think we actually need a compatibility method here, if the only change is an argument going from `int32_t` to `int64_t`. I think it could probably be removed and GDExtension would be fine. If the validator gives a "hash has changed" error, I think that'd be a bug in the validator.